# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Uúba Tech API",
...
    "contact": {
      "name": "Uúba Tech",
      "url": "https://uuba.tech/"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/faturas": {
      "post": {
        "tags": [
          "faturas"
        ],
        "summary": "Registrar fatura",
        "description": "Cria uma nova fatura a receber vinculada a um cliente. O valor deve ser em centavos (R$ 2.500,00 = `250000`).",
        "operationId": "create_fatura_api_v1_faturas_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaturaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaturaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "faturas"
        ],
        "summary": "Listar faturas",
        "description": "Retorna faturas com filtros opcionais por status e cliente. Use `status=pendente,vencido` para faturas em aberto.",
        "operationId": "list_faturas_api_v1_faturas_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por status (aceita múltiplos separados por vírgula: pendente,vencido)",
              "title": "Status"
            },
            "description": "Filtrar por status (aceita múltiplos separados por vírgula: pendente,vencido)"
          },
          {
            "name": "cliente_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por cliente",
              "title": "Cliente Id"
            },
            "description": "Filtrar por cliente"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Itens por página (max 100)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Itens por página (max 100)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pular N itens",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pular N itens"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faturas/{fatura_id}": {
      "get": {
        "tags": [
          "faturas"
        ],
        "summary": "Buscar fatura",
        "description": "Retorna os dados completos de uma fatura pelo ID.",
        "operationId": "get_fatura_api_v1_faturas__fatura_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "fatura_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fatura Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaturaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "faturas"
        ],
        "summary": "Atualizar fatura",
        "description": "Atualiza status, promessa de pagamento, ou link de pagamento. Ao marcar como `pago`, o campo `pago_em` é preenchido automaticamente.",
        "operationId": "update_fatura_api_v1_faturas__fatura_id__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "fatura_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fatura Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaturaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaturaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cobrancas": {
      "post": {
        "tags": [
          "cobrancas"
        ],
        "summary": "Registrar cobrança",
        "description": "Registra uma ação de cobrança vinculada a uma fatura. Especifique o tipo (lembrete, cobrança, follow_up, escalação) e opcionalmente o tom e mensagem.",
        "operationId": "create_cobranca_api_v1_cobrancas_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CobrancaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cobrancas"
        ],
        "summary": "Listar cobranças",
        "description": "Retorna cobranças com filtros por período, cliente, ou fatura. Use `periodo=7d` para cobranças dos últimos 7 dias.",
        "operationId": "list_cobrancas_api_v1_cobrancas_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "periodo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Período em dias (ex: 7d, 30d)",
              "title": "Periodo"
            },
            "description": "Período em dias (ex: 7d, 30d)"
          },
          {
            "name": "cliente_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por cliente",
              "title": "Cliente Id"
            },
            "description": "Filtrar por cliente"
          },
          {
            "name": "fatura_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por fatura",
              "title": "Fatura Id"
            },
            "description": "Filtrar por fatura"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Itens por página (max 100)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Itens por página (max 100)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pular N itens",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pular N itens"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cobrancas/{fatura_id}/historico": {
      "get": {
        "tags": [
          "cobrancas"
        ],
        "summary": "Histórico de cobranças",
        "description": "Retorna toda a timeline de cobranças enviadas para uma fatura específica, ordenadas da mais recente para a mais antiga.",
        "operationId": "get_historico_api_v1_cobrancas__fatura_id__historico_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "fatura_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fatura Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cobrancas/{cobranca_id}/pausar": {
      "patch": {
        "tags": [
          "cobrancas"
        ],
        "summary": "Pausar cobrança",
        "description": "Pausa a régua de cobrança para esta fatura. Nenhuma mensagem será enviada enquanto estiver pausada.",
        "operationId": "pausar_api_v1_cobrancas__cobranca_id__pausar_patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "cobranca_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cobranca Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cobrancas/{cobranca_id}/retomar": {
      "patch": {
        "tags": [
          "cobrancas"
        ],
        "summary": "Retomar cobrança",
        "description": "Retoma uma cobrança que estava pausada. A régua volta a funcionar normalmente.",
        "operationId": "retomar_api_v1_cobrancas__cobranca_id__retomar_patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "cobranca_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cobranca Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CobrancaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CobrancaCreate": {
        "properties": {
          "fatura_id": {
            "type": "string",
            "title": "Fatura Id"
          },
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "lembrete",
              "cobranca",
              "follow_up",
              "escalacao"
            ],
            "title": "Tipo"
          },
          "canal": {
            "type": "string",
            "enum": [
              "whatsapp",
              "email",
              "sms"
            ],
            "title": "Canal",
            "default": "whatsapp"
          },
          "mensagem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mensagem"
          },
          "tom": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "amigavel",
                  "neutro",
                  "firme",
                  "urgente"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tom"
          }
        },
        "type": "object",
        "required": [
          "fatura_id",
          "cliente_id",
          "tipo"
        ],
        "title": "CobrancaCreate"
      },
      "CobrancaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "cobranca"
          },
          "fatura_id": {
            "type": "string",
            "title": "Fatura Id"
          },
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "canal": {
            "type": "string",
            "title": "Canal"
          },
          "mensagem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mensagem"
          },
          "tom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tom"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pausado": {
            "type": "boolean",
            "title": "Pausado"
          },
          "agent_decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Decision Id"
          },
          "enviado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enviado Em"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fatura_id",
          "cliente_id",
          "tipo",
          "canal",
          "status",
          "pausado",
          "created_at",
          "updated_at"
        ],
        "title": "CobrancaResponse"
      },
      "FaturaCreate": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "valor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Valor"
          },
          "vencimento": {
            "type": "string",
            "format": "date-time",
            "title": "Vencimento"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "numero_nf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Nf"
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "valor",
          "vencimento"
        ],
        "title": "FaturaCreate"
      },
      "FaturaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "fatura"
          },
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "valor": {
            "type": "integer",
            "title": "Valor"
          },
          "moeda": {
            "type": "string",
            "title": "Moeda"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "vencimento": {
            "type": "string",
            "format": "date-time",
            "title": "Vencimento"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "numero_nf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Nf"
          },
          "pagamento_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagamento Link"
          },
          "pago_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pago Em"
          },
          "promessa_pagamento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promessa Pagamento"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cliente_id",
          "valor",
          "moeda",
          "status",
          "vencimento",
          "created_at",
          "updated_at"
        ],
        "title": "FaturaResponse"
      },
      "FaturaUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pendente",
                  "pago",
                  "vencido",
                  "cancelado"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "promessa_pagamento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promessa Pagamento"
          },
          "pagamento_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagamento Link"
          }
        },
        "type": "object",
        "title": "FaturaUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list"
          },
          "data": {
            "items": {},
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "ListResponse"
      },
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "total",
          "page_size",
          "has_more"
        ],
        "title": "PaginationMeta"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}